% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.chordNetwork}
\alias{as.chordNetwork}
\title{Convert an R hclust or dendrogram object into a chordNetwork list.}
\usage{
as.chordNetwork(d, classes)
}
\arguments{
\item{d}{An object of R class \code{hclust} or \code{dendrogram}.}

\item{classes}{An optional name for the classes node. If missing, use the first argument
variable name.}
}
\description{
\code{as.chordNetwork} converts an R hclust or dendrogram object into a list suitable
for use by the \code{chordNetwork} function.
}
\details{
\code{as.chordNetwork} coverts R objects of class \code{hclust} or
\code{dendrogram} into a list suitable for use with the \code{chordNetwork} function.
}
\examples{
# Create a hierarchical chord object and display with chordNetwork
## dontrun
hc <- hclust(dist(USArrests), "ave")
chordNetwork(as.chordNetwork(hc))
}

