% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.clusterNetwork}
\alias{as.clusterNetwork}
\title{Convert an R hclust or dendrogram object into a clusterNetwork list.}
\usage{
as.clusterNetwork(d, classes)
}
\arguments{
\item{d}{An object of R class \code{hclust} or \code{dendrogram}.}

\item{classes}{An optional name for the classes node. If missing, use the first argument
variable name.}
}
\description{
\code{as.clusterNetwork} converts an R hclust or dendrogram object into a list suitable
for use by the \code{clusterNetwork} function.
}
\details{
\code{as.clusterNetwork} coverts R objects of class \code{hclust} or
\code{dendrogram} into a list suitable for use with the \code{clusterNetwork} function.
}
\examples{
# Create a hierarchical cluster object and display with clusterNetwork
## dontrun
hc <- hclust(dist(USArrests), "ave")
clusterNetwork(as.clusterNetwork(hc))
}

