% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.d3Tree}
\alias{as.d3Tree}
\title{Convert an R hclust or dendrogram object into a d3Tree list.}
\usage{
as.d3Tree(d, root)
}
\arguments{
\item{d}{An object of R class \code{hclust} or \code{dendrogram}.}

\item{root}{An optional name for the root node. If missing, use the first argument
variable name.}
}
\description{
\code{as.d3Tree} converts an R hclust or dendrogram object into a list suitable
for use by the \code{d3Tree} function.
}
\details{
\code{as.d3Tree} coverts R objects of class \code{hclust} or
\code{dendrogram} into a list suitable for use with the \code{d3Tree} function.
}
\examples{
# Create a hierarchical cluster object and display with d3Tree
## dontrun
hc <- hclust(dist(USArrests), "ave")
d3Tree(as.d3Tree(hc))
}

