% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.treemapNetwork}
\alias{as.treemapNetwork}
\title{Convert an R hclust or dendrogram object into a treemapNetwork list.}
\usage{
as.treemapNetwork(d, classes)
}
\arguments{
\item{d}{An object of R class \code{hclust} or \code{dendrogram}.}

\item{classes}{An optional name for the classes node. If missing, use the first argument
variable name.}
}
\description{
\code{as.treemapNetwork} converts an R hclust or dendrogram object into a list suitable
for use by the \code{treemapNetwork} function.
}
\details{
\code{as.treemapNetwork} coverts R objects of class \code{hclust} or
\code{dendrogram} into a list suitable for use with the \code{treemapNetwork} function.
}
\examples{
# Create a hierarchical treemap object and display with treemapNetwork
## dontrun
hc <- hclust(dist(USArrests), "ave")
treemapNetwork(as.treemapNetwork(hc))
}

