% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.hclust}
\alias{as_treenetdf.hclust}
\title{Convert hclust objects to \code{treenetdf}}
\usage{
\method{as_treenetdf}{hclust}(data, ...)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details).}

\item{...}{arguments to pass to methods.}
}
\description{
Convert hclust objects to \code{treenetdf}
}
