% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.igraph}
\alias{as_treenetdf.igraph}
\title{Convert igraph tree to \code{treenetdf}}
\usage{
\method{as_treenetdf}{igraph}(data, root = "root", ...)
}
\arguments{
\item{data}{an object to convert to \code{treenetdf}.}

\item{root}{character string naming the root in \code{data}.}

\item{...}{arguments to pass to methods.}
}
\description{
Convert igraph tree to \code{treenetdf}
}
