% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.Node}
\alias{as_treenetdf.Node}
\title{data.tree to \code{treenetdf}}
\usage{
\method{as_treenetdf}{Node}(data, ...)
}
\arguments{
\item{data}{an object to convert to \code{treenetdf}.}

\item{...}{arguments to pass to methods.}
}
\description{
data.tree to \code{treenetdf}
}
