% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.phylo}
\alias{as_treenetdf.phylo}
\title{Phylo tree to \code{treenetdf}}
\usage{
\method{as_treenetdf}{phylo}(data, ...)
}
\arguments{
\item{data}{a tree network description in one of numerous forms (see
details).}

\item{...}{arguments to pass to methods.}
}
\description{
Phylo tree to \code{treenetdf}
}
