% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.tbl_graph}
\alias{as_treenetdf.tbl_graph}
\title{tbl_graph_to_treenetdf}
\usage{
\method{as_treenetdf}{tbl_graph}(data, ...)
}
\arguments{
\item{data}{an object to convert to \code{treenetdf}.}

\item{...}{arguments to pass to methods.}
}
\description{
tbl_graph_to_treenetdf
}
