% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{Create a d3 bundle}
\usage{
bundle(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{Json data to use for the bundle}

\item{width}{The width of the window to put the svg in; defaults to automatically expand to fit the svg}

\item{height}{The height of the window to put the svg in; defaults to automatically expand to fit the svg}
}
\description{
Create a d3 bundle
}

