% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundleOutput}
\alias{bundleOutput}
\title{Create a bundle output}
\usage{
bundleOutput(outputId, width = "100\%", height = "1200px")
}
\arguments{
\item{outputId}{The ID that you will use in server.R to render to this output}

\item{width}{The width of the SVG}

\item{height}{The height of the SVG}
}
\description{
Create a bundle output
}

