% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chordNetwork}
\alias{chordNetwork}
\title{Create Chord network diagram.}
\source{
Reingold. E. M., and Tilford, J. S. (1981). Tidier Drawings of Chords.
IEEE Transactions on Software Engineering, SE-7(2), 223-228.

Mike Bostock: \url{http://bl.ocks.org/mbostock/4063550}.
}
\usage{
chordNetwork(matrix, digits = 18, df, height = NULL, width = NULL,
  fontSize = 10, linkColour = "#ccc", nodeColour = "#fff",
  nodeStroke = "steelblue", textColour = "#111", opacity = 0.9,
  margin = 0)
}
\arguments{
\item{height}{height for the network graph's frame area in pixels (if
\code{NULL} then height is automatically determined based on context)}

\item{width}{numeric width for the network graph's frame area in pixels (if
\code{NULL} then width is automatically determined based on context)}

\item{fontSize}{numeric font size in pixels for the node text labels.}

\item{linkColour}{character string specifying the colour you want the link
lines to be. Multiple formats supported (e.g. hexadecimal).}

\item{nodeColour}{character string specifying the colour you want the node
circles to be. Multiple formats supported (e.g. hexadecimal).}

\item{nodeStroke}{character string specifying the colour you want the node
perimeter to be. Multiple formats supported (e.g. hexadecimal).}

\item{textColour}{character string specifying the colour you want the text to
be before they are clicked. Multiple formats supported (e.g. hexadecimal).}

\item{opacity}{numeric value of the proportion opaque you would like the
graph elements to be.}

\item{margin}{integer value of the plot margin. Set the margin
appropriately to accomodate long text labels.}

\item{List}{a hierarchical list object with a root node and children.}
}
\description{
Create Chord network diagram.
}
\examples{
## dontrun
## Create chord from JSON formatted data

table <- read.csv(system.file(file.path("data", "cities.csv"), package = "networkD3"))
matrix <- JSONtoMatrix(file = system.file(file.path("data", "matrix.json"), package = "networkD3"))

# Visualize the chord
chordNetwork(matrix = matrix, digits = 18, df = table, fontSize = 10)
}

