% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{energy}
\alias{energy}
\title{JSON data file of a projection of UK energy production and consumption in 2050.}
\format{A JSON file with two arrays \code{nodes} and \code{links}.}
\source{
See Mike Bostock \url{http://bost.ocks.org/mike/sankey/}.
}
\description{
JSON data file of a projection of UK energy production and consumption in 2050.
}
\keyword{datasets}

