% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{flare}
\alias{flare}
\title{JSON data file of the Flare class hierarchy.}
\format{A JSON file with two arrays \code{name} and \code{children}.}
\source{
See Mike Bostock \url{http://bl.ocks.org/mbostock/4063550}.
}
\description{
JSON data file of the Flare class hierarchy.
}
\keyword{datasets}

