% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flexNetwork.R
\name{flexNetwork}
\alias{flexNetwork}
\title{Create Reingold-Tilford Tree network diagrams with different node sizes.}
\source{
Reingold. E. M., and Tilford, J. S. (1981). Tidier Drawings of Trees.
IEEE Transactions on Software Engineering, SE-7(2), 223-228.

Mike Bostock: \url{http://bl.ocks.org/mbostock/4339083}.
}
\usage{
flexNetwork(List, height = NULL, width = NULL, fontSize = 10,
  fontFamily = "serif", linkColour = "#ccc", nodeColour = "#fff",
  nodeStroke = "steelblue", textColour = "#111", opacity = 0.9,
  margin = NULL)
}
\arguments{
\item{List}{a hierarchical list object with a root node and children.}

\item{height}{height for the network graph's frame area in pixels (if
\code{NULL} then height is automatically determined based on context)}

\item{width}{numeric width for the network graph's frame area in pixels (if
\code{NULL} then width is automatically determined based on context)}

\item{fontSize}{numeric font size in pixels for the node text labels.}

\item{fontFamily}{font family for the node text labels.}

\item{linkColour}{character string specifying the colour you want the link
lines to be. Multiple formats supported (e.g. hexadecimal).}

\item{nodeColour}{character string specifying the colour you want the node
circles to be. Multiple formats supported (e.g. hexadecimal).}

\item{nodeStroke}{character string specifying the colour you want the node
perimeter to be. Multiple formats supported (e.g. hexadecimal).}

\item{textColour}{character string specifying the colour you want the text to
be before they are clicked. Multiple formats supported (e.g. hexadecimal).}

\item{opacity}{numeric value of the proportion opaque you would like the
graph elements to be.}

\item{margin}{an integer or a named \code{list}/\code{vector} of integers
for the plot margins. If using a named \code{list}/\code{vector},
the positions \code{top}, \code{right}, \code{bottom}, \code{left}
are valid.  If a single integer is provided, then the value will be
assigned to the right margin. Set the margin appropriately
to accomodate long text labels.}
}
\description{
Create Reingold-Tilford Tree network diagrams with different node sizes.
}

