% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JSONtoMatrix}
\alias{JSONtoMatrix}
\title{Read a JSON file into R as matrix.}
\usage{
JSONtoMatrix(jsonStr = NULL, file = NULL, simplify = TRUE)
}
\description{
\code{JSONtoMatrix} reads a JSON data file into R and converts it to a matrix.
}

