% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceNetwork.R
\name{labelScaleEffect}
\alias{labelScaleEffect}
\title{Javascript function to scale the label}
\usage{
labelScaleEffect(fontSize = 14, offset = 13, duration = 300,
  opacity = 1)
}
\arguments{
\item{fontSize}{a scale factor of fontSize}

\item{offset}{a value to x-offset label}

\item{duration}{the duration of the effect}

\item{opacity}{numeric value of the proportion opaque you would like the
graph elements to be.}
}
\description{
Javascript function to scale the label
}
