% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{MisLinks}
\alias{MisLinks}
\title{Les Miserables character links}
\format{A data set with 254 observations of 3 variables.}
\source{
See Mike Bostock \url{http://bl.ocks.org/mbostock/4062045}.
}
\usage{
MisLinks
}
\description{
A data file of links from Knuth's Les Miserables characters data base.
}
\keyword{datasets}
