% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{MisNodes}
\alias{MisNodes}
\title{Les Miserables character nodes}
\format{A data set with 77 observations of 2 variables, plus made up
node size variable.}
\source{
See Mike Bostock \url{http://bl.ocks.org/mbostock/4062045}.
}
\usage{
MisNodes
}
\description{
A data file of nodes from Knuth's Les Miserables characters data base.
}
\keyword{datasets}
