% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkD3.R
\name{networkD3-package}
\alias{networkD3-package}
\alias{networkD3}
\alias{_PACKAGE}
\title{Tools for Creating D3 Network Graphs from R}
\description{
Creates D3 JavaScript network, tree, dendrogram, and Sankey graphs from R.
}
