% Generated by roxygen2 (4.0.1): do not edit by hand
\name{networkD3shiny}
\alias{networkD3shiny}
\alias{renderSimpleNetwork}
\alias{simpleNetworkOutput}
\title{Shiny output bindings for networkD3 widgets}
\usage{
simpleNetworkOutput(outputId, width = "100\%", height = "400px")

renderSimpleNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width,height}{Must be a valid CSS unit (like
\code{"100\%"}, \code{"400px"}, \code{"auto"}) or a number, which will be
coerced to a string and have \code{"px"} appended.}

\item{expr}{An expression that generates a networkD3 visualization.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
  is useful if you want to save an expression in a variable.}
}
\description{
Functions used to include networkD3 widgets within Shiny applications.
}

