% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nodes_links}
\alias{nodes_links}
\title{Generate nodes and links dataframe for using with sankeyNetwork function}
\usage{
nodes_links(dataframe, variables)
}
\arguments{
\item{dataframe}{the name of the dataframe}

\item{variables}{a vector of the variables (factors) to display from left to right}
}
\value{
2 dataframe "nodes" and "links" usable in sankeyNetwork function
}
\description{
Generate nodes and links dataframe for using with sankeyNetwork function
}
\examples{
# A simple sankeyNetwork generated with nodes_links

nodes_links(esoph, c("agegp", "alcgp", "tobgp"))
networkD3::sankeyNetwork(Links = links, Nodes = nodes, Source = 'source',
                            Target = 'target', Value = 'value', NodeID = 'name',
                            units = 'TWh', fontSize = 12, nodeWidth = 30)



#A basic shiny app with a sankeyNetwork generated with nodes_links

if (interactive()) {
 
 shiny::shinyApp(
   ui = shiny::fluidPage(
     shiny::sidebarLayout(
       shiny::sidebarPanel(
         shiny::actionButton("newflowchart", "New flow chart")
       ),
       shiny::mainPanel(
         networkD3::sankeyNetworkOutput("plot")
       )
     )
   ),
   server = function(input, output) {
     output$plot <- networkD3::renderSankeyNetwork({
       input$newflowchart
       nodes_links(esoph, c("agegp", "alcgp", "tobgp"))

       networkD3::sankeyNetwork(Links = links, Nodes = nodes, Source = 'source',
                                Target = 'target', Value = 'value', NodeID = 'name',
                               units = 'TWh', fontSize = 12, nodeWidth = 30)
     })
   }
)}
}
