% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceNetwork.R
\name{nodeSizeEffect}
\alias{nodeSizeEffect}
\title{Javascript function to increase node size}
\usage{
nodeSizeEffect(plusSize = 5, duration = 300)
}
\arguments{
\item{plusSize}{value to increase size node}

\item{duration}{the duration of the effect}
}
\description{
Javascript function to increase node size
}
