% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_installed}
\alias{pkg_installed}
\title{Check if a package is installed}
\usage{
pkg_installed(pkg_name)
}
\arguments{
\item{pkg_name}{character string name of package}
}
\description{
Check if a package is installed
}
