% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bundle.R
\name{renderBundle}
\alias{renderBundle}
\title{Render a bundle}
\usage{
renderBundle(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a bundle, probably using this package's bundle() method}

\item{env}{Environment to evaluate the code in.}

\item{quoted}{If we should force quoted or not.}
}
\description{
Render a bundle
}

