% Generated by roxygen2 (4.0.1): do not edit by hand
\name{saveNetwork}
\alias{saveNetwork}
\title{Save a network graph to an HTML file}
\usage{
saveNetwork(network, file, selfcontained = TRUE)
}
\arguments{
\item{network}{Network to save (e.g. result of calling the function
  \code{simpleNetwork}).}
}
\description{
Save a networkD3 graph to an HTML file for sharing with others. The HTML can
include it's dependencies in an adjacent directory or can bundle all
dependencies into the HTML file (via base64 encoding).
}

