% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-definitions.R
\docType{data}
\name{SchoolsJournals}
\alias{SchoolsJournals}
\title{Edge list of REF (2014) journal submissions for Politics and International
Relations}
\format{A data set with 2732 rows and 3 variables.}
\source{
See REF 2014 \url{http://results.ref.ac.uk/DownloadSubmissions/ByUoa/21}.
}
\usage{
SchoolsJournals
}
\description{
Edge list of REF (2014) journal submissions for Politics and International
Relations
}
\keyword{datasets}

