% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleNetworkOutput}
\alias{renderSimpleNetwork}
\alias{simpleNetworkOutput}
\title{Shiny bindings for simpleNetwork}
\usage{
simpleNetworkOutput(id, width = "100\%", height = 400)

renderSimpleNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Shiny bindings for simpleNetwork
}

