% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{treenetdf_to_nestedlist}
\alias{treenetdf_to_nestedlist}
\title{Convert a treenet to a nested list}
\usage{
treenetdf_to_nestedlist(df, id_col = "nodeId", parent_col = "parentId")
}
\arguments{
\item{df}{a \code{treenetdf} to convert to a nested list}

\item{id_col}{character string indicating the node ID in \code{df}}

\item{parent_col}{character string indicating the parent ID in \code{df}}
}
\description{
Convert a treenet to a nested list
}
