% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceNetwork.R
\name{unFocusOtherLinks}
\alias{unFocusOtherLinks}
\alias{unFocusOtherNodes}
\title{Javascript function to unfocus other nodes}
\usage{
unFocusOtherLinks(unfocusDivisor = 2, duration = 200, opacity = 1,
  reset = FALSE)

unFocusOtherNodes(unfocusDivisor = 2, duration = 200, opacity = 1,
  reset = FALSE)
}
\arguments{
\item{unfocusDivisor}{a divisor factor of apacity}

\item{duration}{the duration of the effect}

\item{opacity}{numeric value of the proportion opaque you would like the
graph elements to be.}

\item{reset}{reset the effect}
}
\description{
Javascript function to unfocus other nodes
}
