% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.data.frame}
\alias{as_treenetdf.data.frame}
\title{data.frame_to_treenetdf}
\usage{
\method{as_treenetdf}{data.frame}(data = NULL, cols = setNames(names(data),
  names(data)), dftype = "treenetdf", subset = names(data), root = NULL,
  ...)
}
\description{
data.frame_to_treenetdf
}
