% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf.list}
\alias{as_treenetdf.list}
\title{Convert a nested list to \code{treenetdf}}
\usage{
\method{as_treenetdf}{list}(data, children_name = "children",
  node_name = "name", ...)
}
\arguments{
\item{data}{an object to convert to \code{treenetdf}.}

\item{children_name}{character string identifying the child names in
\code{data}.}

\item{node_name}{character string identifying the node names in
\code{data}.}

\item{...}{arguments to pass to methods.}
}
\description{
Convert a nested list to \code{treenetdf}
}
