% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_treenetdf.R
\name{as_treenetdf}
\alias{as_treenetdf}
\title{Convert an object to a \code{treenetdf}}
\usage{
as_treenetdf(data = NULL, ...)
}
\arguments{
\item{data}{an object to convert to \code{treenetdf}.}
}
\description{
Convert an object to a \code{treenetdf}
}
