% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{JS}
\alias{JS}
\title{Create character strings that will be evaluated as JavaScript}
\source{
A direct import of \code{JS} from Ramnath Vaidyanathan, Yihui Xie,
JJ Allaire, Joe Cheng and Kenton Russell (2015). \link{htmlwidgets}: HTML
Widgets for R. R package version 0.4.
}
\usage{
JS(...)
}
\arguments{
\item{...}{character string to evaluate}
}
\description{
Create character strings that will be evaluated as JavaScript
}

