% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MisLinks}
\alias{MisLinks}
\title{A data file of links from Knuth's Les Miserables characters data base.}
\format{A data set with 254 observations of 3 variables.}
\source{
See Mike Bostock \url{http://bl.ocks.org/mbostock/4062045}.
}
\description{
A data file of links from Knuth's Les Miserables characters data base.
}
\keyword{datasets}

