% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MisNodes}
\alias{MisNodes}
\title{A data file of nodes from Knuth's Les Miserables characters data base.}
\format{A data set with 77 observations of 2 variables.}
\source{
See Mike Bostock \url{http://bl.ocks.org/mbostock/4062045}.
}
\description{
A data file of nodes from Knuth's Les Miserables characters data base.
}
\keyword{datasets}

