% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkD3.R
\docType{package}
\name{networkD3-package}
\alias{networkD3}
\alias{networkD3-package}
\title{Tools for creating D3 network graphs from R}
\description{
This packages is intended to make it easy to create D3 JavaScript force
network and Sankey graphs from R using data frames.
}

