% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{networkD3-package}
\alias{networkD3}
\alias{networkD3-package}
\title{Tools for creating D3 network graphs from R}
\description{
This packages is intended to make it easy to create D3 JavaScript
network, tree, dendrogram, and Sankey graphs from R using data frames.
}

