% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/networkD3.R
\docType{package}
\name{networkD3-package}
\alias{networkD3}
\alias{networkD3-package}
\title{Tools for Creating D3 Network Graphs from R}
\description{
Creates D3 JavaScript network, tree, dendrogram, and Sankey graphs from R.
}

