% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radialNetwork.R
\name{radialNetwork}
\alias{radialNetwork}
\title{Radial Network Diagrams}
\usage{
radialNetwork(List, options = list(), tasks = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{List}{a hierarchical \code{list} object with a root node and children.}

\item{options}{\code{list} of options for the \code{radialNetwork}.}

\item{tasks}{\code{list} of \code{htmlwidgets::JS} functions to perform after the
the radialNetwork is rendered.}

\item{width}{numeric width for the network graph's frame area in pixels (if
  \code{NULL} then width is automatically determined based on context)}

\item{height}{height for the network graph's frame area in pixels (if
\code{NULL} then height is automatically determined based on context)}
}
\description{
Create \code{d3.js} radial network diagrams with helpful interactivity.  These diagrams
are based off the code from this \href{https://gist.github.com/wmleler/a734fb2bb3319a2cb386}{gist}.
}
\examples{
\dontrun{
# Download data
library(RCurl)
Flare <- getURL("https://gist.githubusercontent.com/mbostock/4063550/raw/a05a94858375bd0ae023f6950a2b13fac5127637/flare.json")

# Convert to list format
Flare <- rjson::fromJSON(Flare)

# Create radial network
radialNetwork(Flare)
}
}

