% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/radialNetwork.R
\name{radialNetwork}
\alias{radialNetwork}
\title{Create radial network diagrams.}
\usage{
radialNetwork(List, options = list(), width = NULL, height = NULL)
}
\arguments{
\item{List}{a hierarchical list object with a root node and children.}

\item{width}{numeric width for the network graph's frame area in pixels (if
  \code{NULL} then width is automatically determined based on context)}

\item{height}{height for the network graph's frame area in pixels (if
\code{NULL} then height is automatically determined based on context)}
}
\description{
Create radial network diagrams.
}

