% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeNetwork.R
\name{treeNetwork}
\alias{treeNetwork}
\title{<Add Title>}
\usage{
treeNetwork(data, width = NULL, height = NULL, elementId = NULL,
  margin = NULL, treeType = "tidy", direction = "right",
  linkType = "diagonal", textRotate = 0, css = NULL, defaults = NULL,
  mouseover = "", mouseout = "", inbrowser = FALSE, ...)
}
\description{
<Add Description>
}
