% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treeNetwork.R
\name{treeNetwork}
\alias{treeNetwork}
\title{Inform user that treeNetwork has been depricated.}
\usage{
treeNetwork(...)
}
\arguments{
\item{...}{arguments to pass to function.}
}
\description{
Inform user that treeNetwork has been depricated.
}

